Moved basiclayout components to component folder & added collapsable side menu
