+++ client/src/core/BasicLayout/BasicLayout.jsx
@@ -1,15 +1,18 @@
 import React, { useContext, useEffect } from "react";
-import Menu from "../Menu";
-import SideMenu from "../SideMenu";
+import Menu from "./components/Menu/Menu";
+import SideMenu from "./components/SideMenu/SideMenu";
 import "./BasicLayout.css";
 import "./BasicLayoutStyles.css";
 import "../OverrideDefaultMD.css";
 import { AuthContext } from "../../auth/AuthContext";
 import { Redirect } from "react-router";
 import KUTE from "kute.js";
+import Wave from "./components/Wave";
+import { useState } from "react";
 
 const BasicLayout = ({ children }) => {
   const authContext = useContext(AuthContext);
+  const [isEnabled, setIsEnabled] = useState(false);
   useEffect(() => {
     if (!authContext.isAuthenticated()) return;
 
@@ -35,44 +38,11 @@ const BasicLayout = ({ children }) => {
     return (
       <div className="interface__wrapper">
         <Menu />
-        <SideMenu />
-        <main className="main__wrapper">{children}</main>
-        <div className="bottom-wave">
-          <svg
-            id="visual"
-            viewBox="0 0 960 500"
-            width="960"
-            height="500"
-            xmlns="http://www.w3.org/2000/svg"
-            version="1.1"
-          >
-            <g>
-              <path
-                id="p2"
-                d="M0 423L16 416.2C32 409.3 64 395.7 96 390C128 384.3 160 386.7 192 387C224 387.3 256 385.7 288 384.7C320 383.7 352 383.3 384 389.3C416 395.3 448 407.7 480 408.3C512 409 544 398 576 394C608 390 640 393 672 397.5C704 402 736 408 768 410.7C800 413.3 832 412.7 864 407.7C896 402.7 928 393.3 944 388.7L960 384L960 541L944 541C928 541 896 541 864 541C832 541 800 541 768 541C736 541 704 541 672 541C640 541 608 541 576 541C544 541 512 541 480 541C448 541 416 541 384 541C352 541 320 541 288 541C256 541 224 541 192 541C160 541 128 541 96 541C64 541 32 541 16 541L0 541Z"
-                fill="#24242b"
-              ></path>
-              <path
-                id="p3"
-                d="M0 454L16 454.7C32 455.3 64 456.7 96 457.7C128 458.7 160 459.3 192 464.3C224 469.3 256 478.7 288 483.3C320 488 352 488 384 486C416 484 448 480 480 474.7C512 469.3 544 462.7 576 465.8C608 469 640 482 672 487.7C704 493.3 736 491.7 768 489.5C800 487.3 832 484.7 864 483.5C896 482.3 928 482.7 944 482.8L960 483L960 541L944 541C928 541 896 541 864 541C832 541 800 541 768 541C736 541 704 541 672 541C640 541 608 541 576 541C544 541 512 541 480 541C448 541 416 541 384 541C352 541 320 541 288 541C256 541 224 541 192 541C160 541 128 541 96 541C64 541 32 541 16 541L0 541Z"
-                fill="#18181d"
-              ></path>
-
-              <path
-                id="pp2"
-                style={{ visibility: "hidden" }}
-                d="M0 417L11.5 418.2C23 419.3 46 421.7 68.8 418.8C91.7 416 114.3 408 137.2 401.7C160 395.3 183 390.7 205.8 388.8C228.7 387 251.3 388 274.2 394.5C297 401 320 413 342.8 419.8C365.7 426.7 388.3 428.3 411.2 429.5C434 430.7 457 431.3 480 426.7C503 422 526 412 548.8 405.7C571.7 399.3 594.3 396.7 617.2 394.2C640 391.7 663 389.3 685.8 392.5C708.7 395.7 731.3 404.3 754.2 406.5C777 408.7 800 404.3 822.8 404.7C845.7 405 868.3 410 891.2 416.5C914 423 937 431 948.5 435L960 439L960 541L948.5 541C937 541 914 541 891.2 541C868.3 541 845.7 541 822.8 541C800 541 777 541 754.2 541C731.3 541 708.7 541 685.8 541C663 541 640 541 617.2 541C594.3 541 571.7 541 548.8 541C526 541 503 541 480 541C457 541 434 541 411.2 541C388.3 541 365.7 541 342.8 541C320 541 297 541 274.2 541C251.3 541 228.7 541 205.8 541C183 541 160 541 137.2 541C114.3 541 91.7 541 68.8 541C46 541 23 541 11.5 541L0 541Z"
-                fill="#24242b"
-              ></path>
-              <path
-                id="pp3"
-                style={{ visibility: "hidden" }}
-                d="M0 463L11.5 461C23 459 46 455 68.8 460.5C91.7 466 114.3 481 137.2 480C160 479 183 462 205.8 459C228.7 456 251.3 467 274.2 467.8C297 468.7 320 459.3 342.8 453.8C365.7 448.3 388.3 446.7 411.2 446C434 445.3 457 445.7 480 446.7C503 447.7 526 449.3 548.8 454.2C571.7 459 594.3 467 617.2 473.2C640 479.3 663 483.7 685.8 480.2C708.7 476.7 731.3 465.3 754.2 462.7C777 460 800 466 822.8 464.7C845.7 463.3 868.3 454.7 891.2 450.7C914 446.7 937 447.3 948.5 447.7L960 448L960 541L948.5 541C937 541 914 541 891.2 541C868.3 541 845.7 541 822.8 541C800 541 777 541 754.2 541C731.3 541 708.7 541 685.8 541C663 541 640 541 617.2 541C594.3 541 571.7 541 548.8 541C526 541 503 541 480 541C457 541 434 541 411.2 541C388.3 541 365.7 541 342.8 541C320 541 297 541 274.2 541C251.3 541 228.7 541 205.8 541C183 541 160 541 137.2 541C114.3 541 91.7 541 68.8 541C46 541 23 541 11.5 541L0 541Z"
-                fill="#18181d"
-              ></path>
-            </g>
-          </svg>
-        </div>
+        <SideMenu isEnabled={isEnabled} setIsEnabled={setIsEnabled} />
+        <main className={`main__wrapper ${isEnabled ? "" : "expanded"}`}>
+          {children}
+        </main>
+        <Wave />
       </div>
     );
   else return <Redirect to="/auth/login" />;
