+++ client/src/core/BasicLayout/components/SideMenu/SideMenu.css
@@ -1,3 +1,20 @@
+.nav-side-expander {
+  position: absolute;
+  left: 250px;
+  top: calc(50vh);
+  width: 10px;
+  height: 50px;
+  background-color: rgba(255, 255, 255, 0.3);
+  border-radius: 10px;
+  cursor: pointer;
+  z-index: 999;
+  transition: 300ms;
+}
+
+.nav-side-expander:hover {
+  background-color: rgba(255, 255, 255);
+}
+
 .nav-side {
   position: absolute;
   top: 64px;
@@ -7,8 +24,12 @@
   height: Calc(100vh - 64px);
   background-color: var(--Shark);
   z-index: 999;
-  /* box-shadow: 10px 5px 6px #1d1d22; */
   overflow: hidden;
+  transition: 300ms;
+}
+
+.collapse {
+  transform: translateX(-240px);
 }
 
 .nav-side__container {
@@ -69,9 +90,6 @@
   margin-bottom: 1rem;
   transition: 250ms ease-in-out;
   cursor: pointer;
-  /* display: flex;
-    flex-direction: row;
-    justify-content: space-between; */
   letter-spacing: 0.5px;
   color: var(--Gold);
 
@@ -122,6 +140,10 @@
 
 @media screen and (max-width: 768px) {
   .nav-side {
+    display: none;
+  }
+
+  .nav-side-expander {
     display: none;
   }
 }
